chore: separate style sheets for components
