+++ src/components/NavBar/_index.scss
@@ -1,0 +1,14 @@
+.nav-links {
+  display: flex;
+  .link {
+    list-style-type: none;
+    color: #73cff4;
+    padding: 1vh;
+    align-items: center;
+
+    &:hover,
+    &:focus {
+      transform: scale(1.1);
+    }
+  }
+}
