+++ src/components/PreviewItem/_index.scss
