+++ src/core.scss
@@ -1,9 +1,12 @@
+@use './components/NavBar';
+@use './components/PreviewItem';
+
 body {
   display: flex;
   flex-direction: column;
 
   margin: 0px;
-  height:100vh;
+  height: 100vh;
 
   background-color: #090f17;
   color: #73cff4;
@@ -11,13 +14,14 @@ body {
   font-size: 1.5em;
 }
 
-header {
+.root-header {
+  display: flex;
+  flex-direction: row;
   padding: 2vh 2.5vh 1.5vh;
   background-color: #1d2c49;
-}
-
-h1 {
-  text-align: center;
+  .heading {
+    margin-right: auto;
+  }
 }
 
 p {
@@ -32,26 +36,7 @@ a {
   text-decoration: none;
 }
 
-a:hover, a:focus {
+a:hover,
+a:focus {
   text-decoration: underline;
-}
-
-.nav {
-  display: flex;
-}
-
-.nav-item {
-  list-style-type: none;
-  color: #73cff4;
-  padding: 1vh;
-  align-items: center;
-}
-
-.nav-item:first-child {
-  list-style-type: none;
-  margin-right: auto;
-}
-
-.nav-item:hover, .nav-item:focus {
-  transform: scale(1.1);
 }
